Initialize this repository
The initial commit of this Git repository. The other git repository
(DictationTool-Old) is messed up by Travis CI (full of Travis CI
updating commits). So I decided to create this new repository to
replace it. By the way, I'll never use Travis CI from now on
because I think it's not useful for this project now.

Let it rip!

Signed-off-by: xianfengting <xianfengting221@163.com>